- Added a test for the MappingService.php - Some work on the municipalities - Renamed some tables so the order upon a fresh migrate is ok - Deleted some migrations which were responsible for adding columns, merged those in the original create migrations. - Deleted some migrations which just had no purpose anymore

+++ database/migrations/2023_02_10_103937_add_municipality_id_to_buildings_table.php
@@ -13,10 +13,12 @@ class AddMunicipalityIdToBuildingsTable extends Migration
      */
     public function up()
     {
-        Schema::table('buildings', function (Blueprint $table) {
-            $table->unsignedBigInteger('municipality_id')->after('user_id')->nullable()->default(null);
-            $table->foreign('municipality_id')->references('id')->on('municipalities');
-        });
+        if ( ! Schema::hasColumn('buildings', 'municipality_id')) {
+            Schema::table('buildings', function (Blueprint $table) {
+                $table->unsignedBigInteger('municipality_id')->after('user_id')->nullable()->default(null);
+                $table->foreign('municipality_id')->references('id')->on('municipalities');
+            });
+        }
     }
 
     /**
@@ -26,9 +28,11 @@ class AddMunicipalityIdToBuildingsTable extends Migration
      */
     public function down()
     {
-        Schema::table('buildings', function (Blueprint $table) {
-            $table->dropForeign(['municipality_id']);
-            $table->dropColumn('municipality_id');
-        });
+        if (Schema::hasColumn('buildings', 'municipality_id')) {
+            Schema::table('buildings', function (Blueprint $table) {
+                $table->dropForeign(['municipality_id']);
+                $table->dropColumn('municipality_id');
+            });
+        }
     }
 }
